Changed includes due to OpmLog moved from parser to common
